+++ scrapper/src/main/java/ru/tinkoff/edu/java/scrapper/service/StackOverflowQuestionsService.java
@@ -1,9 +1,11 @@
-package ru.tinkoff.edu.java.scrapper.service.utils;
+package ru.tinkoff.edu.java.scrapper.service;
 
+import org.apache.commons.lang3.ObjectUtils;
 import ru.tinkoff.edu.java.scrapper.client.StackExchangeClient;
 import ru.tinkoff.edu.java.scrapper.client.StackOverflowClient;
 import ru.tinkoff.edu.java.scrapper.configuration.ApplicationConfig;
 import ru.tinkoff.edu.java.scrapper.dto.*;
+import ru.tinkoff.edu.java.scrapper.service.utils.LinkUpdateUtils;
 
 import java.util.ArrayList;
 import java.util.Comparator;
@@ -11,8 +13,30 @@ import java.util.List;
 import java.util.Optional;
 import java.util.function.Function;
 
-public class StackOverflowAnswersUtils {
-    public static <T> List<T> getBatchedUpdates(
+public abstract class StackOverflowQuestionsService {
+    protected List<LinkUpdate> getUpdates(
+            List<StackOverflowQuestion> questions,
+            StackOverflowClient stackOverflowClient,
+            ApplicationConfig applicationConfig,
+            Function<StackOverflowQuestion, List<LinkWithChatId>> getLinks
+    ) {
+        var fetchedBatch = stackOverflowClient.getStackOverflowQuestions(
+                applicationConfig.webClient().stackExchange().apiVersion(),
+                questions.stream().map(StackOverflowQuestion::questionId).toList()
+        );
+        return LinkUpdateUtils.getUpdates(
+                questions,
+                question -> fetchedBatch.items().stream()
+                        .filter(item -> item.questionId().equals(question.questionId()))
+                        .findFirst()
+                        .map(sq -> ObjectUtils.max(sq.lastActivityDate(), sq.lastEditDate()))
+                        .orElse(null),
+                getLinks,
+                StackOverflowQuestion::updatedAt,
+                StackOverflowQuestion::createdAt
+        );
+    }
+    protected  <T> List<T> getBatchedUpdates(
             List<StackOverflowQuestion> questions,
             Function<List<StackOverflowQuestion>, List<T>> getUpdates
     ) {
@@ -25,17 +49,17 @@ public class StackOverflowAnswersUtils {
         return result;
     }
 
-    public static List<StackOverflowAnswerUpdate> getAnswerUpdates(
+    protected List<StackOverflowAnswerUpdate> getAnswerUpdates(
             List<StackOverflowQuestion> questions,
             Function<StackOverflowQuestion, List<LinkWithChatId>> getLinks,
             StackOverflowClient stackOverflowClient,
             ApplicationConfig applicationConfig
     ) {
-        var answers = StackOverflowAnswersUtils.getAnswers(questions, stackOverflowClient, applicationConfig);
+        var answers = getAnswers(questions, stackOverflowClient, applicationConfig);
         return answers
                 .stream()
                 .map(answer ->
-                        StackOverflowAnswersUtils.getAnswerUpdate(
+                        getAnswerUpdate(
                                 answer,
                                 questions,
                                 getLinks))
@@ -44,7 +68,7 @@ public class StackOverflowAnswersUtils {
                 .toList();
     }
 
-    private static List<StackExchangeAnswerResponse> getAnswers(
+    protected List<StackExchangeAnswerResponse> getAnswers(
             List<StackOverflowQuestion> questions,
             StackOverflowClient stackOverflowClient,
             ApplicationConfig applicationConfig
@@ -71,7 +95,7 @@ public class StackOverflowAnswersUtils {
         return answers;
     }
 
-    private static Optional<StackOverflowAnswerUpdate> getAnswerUpdate(
+    protected Optional<StackOverflowAnswerUpdate> getAnswerUpdate(
             StackExchangeAnswerResponse answer,
             List<StackOverflowQuestion> questions,
             Function<StackOverflowQuestion, List<LinkWithChatId>> getLinks
