+++ scrapper/src/main/java/ru/tinkoff/edu/java/scrapper/service/jdbc/JdbcStackOverflowQuestionsService.java
@@ -14,8 +14,7 @@ import ru.tinkoff.edu.java.scrapper.repository.jdbc.JdbcLinksRepository;
 import ru.tinkoff.edu.java.scrapper.repository.jdbc.JdbcStackOverflowQuestionsRepository;
 import ru.tinkoff.edu.java.scrapper.service.FindOrDoService;
 import ru.tinkoff.edu.java.scrapper.service.StackOverflowAnswersService;
-import ru.tinkoff.edu.java.scrapper.service.utils.StackOverflowAnswersUtils;
-import ru.tinkoff.edu.java.scrapper.service.utils.StackOverflowLinksUtils;
+import ru.tinkoff.edu.java.scrapper.service.StackOverflowQuestionsService;
 
 import java.time.OffsetDateTime;
 import java.util.List;
@@ -23,8 +22,9 @@ import java.util.Optional;
 
 @Service
 @RequiredArgsConstructor
-class JdbcStackOverflowQuestionsService implements
-        FindOrDoService<StackOverflowQuestion, StackOverflowParserResult>,
+class JdbcStackOverflowQuestionsService
+        extends StackOverflowQuestionsService
+        implements FindOrDoService<StackOverflowQuestion, StackOverflowParserResult>,
         StackOverflowAnswersService {
     private final JdbcStackOverflowQuestionsRepository jdbcStackOverflowQuestionsRepository;
     private final JdbcLinksRepository jdbcLinksRepository;
@@ -68,9 +68,9 @@ class JdbcStackOverflowQuestionsService implements
 
     @Override
     public List<LinkUpdate> getLinkUpdates(List<StackOverflowQuestion> questions) {
-        return StackOverflowAnswersUtils.getBatchedUpdates(
+        return getBatchedUpdates(
                 questions,
-                batch -> StackOverflowLinksUtils.getUpdates(
+                batch -> getUpdates(
                         batch,
                         stackOverflowClient,
                         applicationConfig,
@@ -93,9 +93,9 @@ class JdbcStackOverflowQuestionsService implements
 
     @Override
     public List<StackOverflowAnswerUpdate> getStackOverflowAnswerUpdates(List<StackOverflowQuestion> questions) {
-        return StackOverflowAnswersUtils.getBatchedUpdates(
+        return getBatchedUpdates(
                 questions,
-                batch -> StackOverflowAnswersUtils.getAnswerUpdates(
+                batch -> getAnswerUpdates(
                         batch,
                         jdbcLinksRepository::findAllWithChatId,
                         stackOverflowClient,
